backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
alembic==1.13.1
pymysql==1.1.0
python-dotenv==1.0.1
apscheduler==3.10.4
httpx==0.27.0
pydantic==2.7.1
PyJWT==2.9.0
